requirements.txt
numpy
scikit-image
scikit-learn
joblib
streamlit
pillow
tqdm
opencv-python-headless
tensorflow==2.15.0